refactor: group paragraphs and blocks in seperate dirs
