+++ packages/next/src/index.js
@@ -1,14 +1,7 @@
 export * from "./app-handler"
 export * from "./node-handler"
-export * from "./paragraphs.container"
-export * from "./paragraphs.controller"
-export * from "./pargaraphs.multiple"
-export * from "./pargaraphs.multiple.tab"
-export * from "./pargaraphs.multiple.accordion"
-export * from "./pargaraphs.template"
-export * from "./banners.template"
-export * from "./blocks.controller"
-export * from "./blocks.template"
+export * from "./paragraphs"
+export * from "./layout-blocks"
 export * from "./api-handler"
 export * from "./utils"
 export * from "./get-translations"
