+++ packages/next/src/layout-blocks/banners.template.jsx
