+++ packages/next/src/layout-blocks/blocks.template.jsx
