+++ packages/next/src/layout-blocks/index.js
@@ -1,0 +1,3 @@
+export * from "./banners.template"
+export * from "./blocks.controller"
+export * from "./blocks.template"
