+++ packages/next/src/menus.jsx
@@ -46,7 +46,6 @@ const fetchMenu = async ({ name, locale }) => {
   }
 
   const data = await response.json()
-  console.log(`fetching ${name}`, data)
   return data
 }
 
@@ -62,7 +61,6 @@ export const getMenu = async (name, locale) => {
 }
 
 export const getMenus = async (names = [], locale) => {
-  console.log("getMenus", locale)
   const requests = names.map((name) =>
     fetchMenu({ name, locale }).then((res) => formatMenu(name, res))
   )
