+++ packages/next/src/paragraphs/index.js
@@ -1,0 +1,7 @@
+export * from "./paragraphs.controller"
+export * from "./paragraphs.container"
+export * from "./paragraphs.controller"
+export * from "./pargaraphs.multiple"
+export * from "./pargaraphs.multiple.tab"
+export * from "./pargaraphs.multiple.accordion"
+export * from "./pargaraphs.template"
