+++ packages/next/src/paragraphs/pargaraphs.multiple.accordion.jsx
